// apps/web/src/types/generated/InitialAllocationCategory.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { InitialAllocationKind } from "./InitialAllocationKind";

export type InitialAllocationCategory = {
  kind: InitialAllocationKind;
  shares: number;
  account_count: number;
  zero_holders: number;
};